Checking 2 scripts to generate SWIG docs from SITK
doxy2swig.py is a modified version of the one in
/playpen/tmp/SimpleITK-build/ITK/Wrapping/Generators/Doc
I changed it to also produce JavaDoc output.

doxyall.py is a script that produces the .i file for
all of SimpleITK.  It can either produce Java or Python-style SWIG files.
The script is run to updated either JavaDoc.i or PythonDocstrings.i
in SimpleITK/Wrapping.  Run the script to see the command line options.