WIP: Read source tree .git directory more robustly
The "GIT_DIR" may not be at ${CMAKE_SOURCE_DIR}/.git and/or may be a
"gitfile".  Use "git rev-parse --git-dir" to look up the directory
location.

If "git gc" packs refs then the ".git/refs/heads/<head>" file may not
exist.  Instead of failing in this case, simply drop the dependency.

This has been tested to work on Windows, but the follow is noted: On
Windows the output may not be a standard Windows

Change-Id: Ib29a2c0cee41254cda1c61e53e089305c69a9bd7