Added support for seriesIDs in querying DICOM
Two new methods were added, a variant of GetGDCMSeriesFileNames and
GetGDCMSeriesIDs.

The new GetGDCMSeriesFileNames has an additional parameter vs. the
previous GetGDCMSeriesFileNames method.  It now has a seriesID string
that the caller supplies to request a specific series if a DICOM
data set has multiple series.

The GetGDCMSeriesIDs method will return a vector of strings containing
all the series in a DICOM data set.

Change-Id: Ide041a20a43a185ff42654066f3794a1355bd2fc