Added a DicomSeriesReader test
The DicomSeriesReader tests ImageSeriesReader's Dicom methods
by calling GetGDCMSeriesIDs and GetGDMSeriesFileNames.
A directory with a 3 image series is scanned for Dicom files.
The series is loaded and compared to a hash value.  Then
The series names are also queried given the specific series ID.

Also the file IOTests.cxx has been moved to sitkImageIOTests.cxx

Change-Id: I3df5f5b2f13c59baff4bbece55d2003fa70ffb99