Removed the "in" prefix from function arguments.
The lua template code generation system has been updated to use the
name of the variable with the first letter lower case, instread of
prefixing with "in".

This has primarily been done for the Python interface since python
allows arguments to be specified by name. However, this causes an
incompatibility with the only argument names.