Creating itkSliceImageFilter for implemented slice notations
The new SliceImageFilter is for assistance with implementing commonly
used slice indexing where a start,stop and step are used. And include
such features as automatically clamping start and stop, supporting
returning and empty image, and negative step to indicate reversing the
direction of a dimension. What is not supported is negative start and
stop. In the ITK realm these are valid indexes, so the common notation
of index from then end has to be implemented at a higher level.

Change-Id: I352d37c2485754f8710d6fd7a8f323ad476da515