ENH: Extract version information from git
The number of commits since the last version changes is now extracted
from the git repository, and added to the version variable as either a
postN for devN suffix. Additionally, the 5 digit SHA1 hash of the git
HEAD is also appended.

As a consequence, a strict protocol needs to be followed when
Version.cmake is changed as well as the formatting and timing of
version tags. Its is expected that tag for a version is made, then the
next commit on the master branch modifies the Version.cmake to
indicate the version development is working towards.

The internal CMake variable to indicate SimpleITK's version have been
renamed to the conventional PROJECT_VERSION_[MAJOR|MINOR...]. Also to
follow standard conventions the SimpleITK_VERSION cmake variable is
defined and the amalgamation of all version information.

TODO: During configuration a CMake file should be generated to store
the extracted git revision information. If there is no git repository
detected then this file should be checked for. When source packaging
happens this file should be included.

Change-Id: I95e5d8876cd58b869cd939fab28d66708fbb9986