Adding missing quotes around @s, and numpy not required
Having a space in the source name caused problems with the
installation due to unquoted @ substitutions.

If the python installation did not have numpy, a failing error would
be returned causing the build to appear to have failed. Now the
installation is just skip. So just a couple tests will fail.

Change-Id: Ia1bca43b0a642b0229f0079660e5be7292c3df6d