Include C99 directory when no stdint.h header it found.
These files also get installed into the SimpleITK directory when not
found.

Because this header file is not wrapped, and has the same name as the
system. It makes code cleaner to maintain the expected name. However,
it may cause conflict and incompatibilities if other project follow
the same pattern. The may be a rare occurrence and at the moment does
not appear worth the change.

Change-Id: I40c83cc09f07101a13c99735a5593fbbb39400cb