Removed exception on Windows for finding ImageJ
Previously FindApplication would throw an exception on Windows if
it could not find ImageJ.  Now it will simply return an empty
string.  Thus if the user specifies a different image viewer
via the SITK_SHOW_COMMAND environment variable, execution can
continue.

If ImageJ is not found and there is no SITK_SHOW_COMMAND given,
an exception is thrown later on.

Change-Id: Icd7570ba1a626647897113d65bb33f01f957ed1b