Adding Python support for SetPixel for Vector and Complex image types
Copied code from __getitem__ to __setitem__ to perform bounds and
error checking in python and to produce more friendly exceptions. The
behavior should be the same, except the index element types in python
must be int, it's possible float or long types were used
before. __setitem__ now call SetPixel to follow the whats done in
GetPixel.

Change-Id: Ic97becbaa640da9b5f7148341643b1d6983fb496