Disable GTest using tr1 tuple
With the current version of Visual Studio 2012 ( VC11 ), tuples are
only instantiated to 5 template parameters which is in sufficient for
gtest. While define _VARIADIC_MAX=10 could improve the situation, this
is also a hack. Since these feature of GTest are not needed we are
just disabling GTest using tuples and hoping that upstream with
permanent fix.

Change-Id: I217848d70110cadc3bfa19f69fa3a41e6a2e8b29