BUG: Change python __invert__ to bitwise not
Compatibility: This changes the behavior of the operator(~) from a
logical not to a bitwise not.

The ITK documentation on these (&^|) operators was incorrectly calling them
"logical" when the were bitwise, but the Not filter was acutally a
logical. So a new bitwise not functor and SimpleITK filter was created
and is not utilized for python.

Change-Id: Ifa5943f488fd40bb05977bfd6356417266ce47b9