BUG: Patch Swig's container class for C++11 from Swig upstream
From upstream swig 92128eef445f75f674894e3f5d4e1fc2a181895:

C++11 support for new versions of erase and insert in the STL containers.

The erase and insert methods in the containers use const_iterator
instead
of iterator in C++11.  There are times when the methods wrapped must
match
the parameters exactly. Specifically when full type information for
template types is missing or SWIG fails to look up the type correctly,
for example:

  %include <std_vector.i>
  typedef float Real;
  %template(RealVector) std::vector<Real>;

SWIG does not find std::vector<Real>::iterator because %template using
typedefs does not always work and so SWIG doesn't know if the type is
copyable and so uses SwigValueWrapper<iterator> which does
not support conversion to another type (const_iterator). This resulted
in compilation errors when using the C++11 version of the containers.