Terminate program is deleting command during an active process
There is a corner case where a user could delete a Command in while a
process is active. This would have the ITK object holding a command
with an invalid sitk Command. There is not easy way to rectify this
situation with out extensive book keeping. So there program is
explicitly terminated.

Change-Id: I0891ca1d2d87f10c28bebc7333d755f67d44fccc