ENH: Make C++11 try_compile result related to CMAKE_CXX_FLAGS
If the CMAKE_CXX_FLAGS change then the C++11 try compile will be
re-done. This is done by computing the MD5 checksum of the flags, and
using that as part of the variable name in the cache.

Also changing variables to lower case, and removing extra CMAKE_ save
variable in function.

Change-Id: I61bb662e8bb1550f9db1960d872d064053ecc9d8