Reduce LabelGeometry instantiation by restricting intensity to real
The combination of images instantiated was too many for the
LabelGeometry image filter. This was causing many system to run out of
memory during compilation of this filter, and way the largest object
filter, and took the longest to compile.

This filter has been reduced to only take real images for the
intensity input.

Change-Id: I3264f7f73c96852cb994c895284a657b270b06ee