ENH: Remove command from active process when deleted
Change the behavior for program termination to simply removing the
command from the ITK process object.

Previously, in the case when there was an active process and the
SimpleITK command was deleted program termination would occur. With
the added book keeping, the specific command(s) can be removed from
the ITK process object.

Change-Id: I590dd75a7826044f4c444dfaf51954342429b348