BUG: Add check for correct size of parameters to transforms
ITK does not always check the size for the fixed parameters and the
parameters passed. This can place the class into an undefined state.

The SimpleITK wrapper not checks that the size is atleast the size
expected of ITK, and performs a shallow copy into an ITK array object
so the size then correctly matches the requirements.

Change-Id: I83b55c54bb54067e6b27f5e0844f8211ba7ccfd0