ENH: Add Generation of additional Execute methods with optional inputs
Add code generation for the public Execute methods to generate an
additional version with the optional inputs.

The internal execute method now takes points to the inputs so that
unspecified inputs can be passes as a NULL pointer. This approach was
choose since it does not add additional methods. It is also safe sine
the public methods still get references and hold onto the object.

Change-Id: I1d4818f1a9792b997b87a5d12b3642f936cb750a