COMP: Fix operator<< for std:vector in custom gtest predicate
The following error message was produced with clang 5.1:
gtest.h:6785:7: error: call to function 'operator<<' that is neither
visible in the template definition nor found by argument-dependent lookup

The error also went on to say that the overloaded operator should be
defined before usage. Therefor the definition of this overloaded
operator has been moved before the gtest include.

Change-Id: I3767891a44aa9bda1a75aadab9300fa5cd8eba1b