DOC: Fixed spacing in the XML to JSON script
GenerateDocumentation.groovy has been updated to add a leading
space for 'ulink' nodes and a trailing space for 'ulink' and
'ref' nodes.  In the former case, it checks if a space precedes
it.  The latter case can lead to an extraneous space, if the
node is immediately followed by a period or other punctuation.

Change-Id: I279d3a6d73a8754abe8baf855c4f00f0a543b0e8