ITKv4.6 changes the results, renaming old baseline
A new feature to this filter was added, which preserves the geometric
and coordinates of the output image. This requires the baseline to be
updated (next patch). Here we rename the old baseline with the ".1" so
it can be used as an alternative with support is added to the test
infrastructure.

Change-Id: I66b5c8b5f1cd84262db3402904742bd370af9951