Implementing Setting of smoothing parameters and interpolators
This interpolator is simply created from the valid interpolator
enums.

The smoothing parameters are handled in a new way. Because different
smoothing algorithm require different ITK displacement filed types,
when the smoothing parameter is set and is different, the methods
create a new ITK DisplacementFieldTransform and sets the new
displacement field to the old one, and then sets the new smoothing
parameters.

Change-Id: If5f60739b6ffb7a74ddbb61809fb29247a6357e2