Changed useSeriesDetails default to false
For ImageSeriesReader::GetGDCMSeriesFileNames the default value of
useSeriesDetails was changed from true to false.  It was cause
the problems when trying to load a specific Dicom Series, i.e.
when the seriesID parameter was specified.

Also added a new example, DicomSeriesReader2.py, which scans a directory
for all Dicom series, and can output a specific, named series.

Change-Id: I1b32cd8ddf82cb4c5ad8a028079f8a19ba385f1d