Change MakeUnique's access-specifier to public
There are several cases there it is needed to make sure that a deep
copy of an object has happened. Making this method public does expose
the user a little to this level of complexity it is needed at times,
and it is better to use this explicit method then an ad-hoc one which
changed the object to implicit make it unique.

Change-Id: I635c2f914093101e8fa0a8d5f4e5bf27374c30c9