Improve ImageRegistrationMethod::SetInitialRegistration copy handling
To be able to observe the progress of registration, the transform
optimization needs to be done "inplace". To preserve const correctness
a const argument method was added which does deep copy of the
transform and then is make as inplace. The does create an additional
copy during initialization, but allows the optimization to be
observable. Additionally, when inplace is explicitly set to false, the
transform is not copied, and will not be observable. This may be
useful for displacement fields, when there are too many parameters to
look at.

Change-Id: Id18c7c90e526987da176538ee19af279e2ddd193