ENH: Add explicit instantiation of common ITK classes
Added CMake option SITK_EXPLICIT_INSTANTIATION to enable explicit
instantiation of ITK classes.

A new library was created which has to be always be static for correct
linkage which contains the instantiation of the ITK templates.

The class choose by examining all the object files with nm, and
counting the number of occurrences of each symbol. The most frequent
classes were choose to be instantiated. A script provided by Hans
Johnson was utilized, a refined version for SimpleITK should be
checked in later. The template argument type which changed to make
them more portable, which includes using itk::SizeValueType, and also
using the long long type.

Change-Id: I80f585d80002dd75ba79f1617360183ec79634b5