Ignore SITKDLL since the explicit library is allows static
Since mixing export specification of the class declaration and the
explicit instantiation does not appear to be portable working, the
explicit library is always build static, and should not have any
export specification when the rest of SimpleITK is built as
shared. This resolve compilation on windows with shared libraries and
the explicit library static.

Change-Id: I221be671f414cacdf24e3bdda8a9740b8b86f211