Use Explicit library by default, always static on windows
Using the Explicit library as static always works when the symbols are
not exported or imported with MSVC and gcc/clang. With the gcc linker
and non-explicit exports using shared libraries also works, although
there are some warning.

If shared explicit library is used with MSVC, there are linking errors
do to symbols in ITK's static libraries conflicting with the export
library. So the explicit library is always static on windows.

Change-Id: I75dba22d5ad2a4a28004ce80d2824f2d8c3373d8