Adding Transform constructor which does try casts to determine type
Previously, the itk::Transform may have been converted into a
itk::Composite transform to be placed in the SimpleITK Pimple, this
not does allow interface casting to the derived transform class. This
patch utilizes a type list visitor to perform a series of
dynamic_casts and typeid check to convert the TransformBase argument
to the appropriate type to be placed into a SimpleITK TransformPimple.

Change-Id: Ie0830534c8d6a21156fdbb3a2ee2baf2fe1531f5