ADD SITK_PYTHON_USE_VIRTUAL_ENV option
Make using python virtual environment optional, by adding support for
using the system python for testing. The PATHONPATH is automatically
added for testing. Using the virtual env requires that
SITK_FORBID_DOWNLOADS is not enables as downloads may be required for
installing required packages, as a result the "dist.Python" target
used for packaging wheels and eggs is not created when forbid
downloads is enabled.

Change-Id: Iaf3a1aef1ad77b04be34e0b06cb3928632e1459d