Add Lua Superbuild support
Download Lua source ball form the midas server.

Create minimal custom CMake file to copy into the Lua build directory
to perform configuration, compilation step and installation of the
executable.

The created binary is passed to the actual SimpleITK project. Or the
system Lua interpreter could be used.

It is not supported compiling SimpleITKLua from the Superbuild Lua.

Change-Id: I335184e249a1362210660001e4aa26422e71fdac