Fix finding lua executable when not explicitly set.
When the SITK_LUA_EXECUTABLE was not explicitly set, the executable
was not search for.

Remove the variable evaluation, to allow CMake to determine if the
variable exists. This allows the initial value of SITK_LUA_EXECUTABLE
to be set, when not explicitly set ( as the superbuild does ).

Remove unconditional find to preserve the lua version used for the
target for wrapping.

Change-Id: I7040fecd8f78db020d8a6ce688fcfc62ca6b54c5