Rewrite Show to support Fiji
Now the Show function will first search for Fiji.  If it finds Fiji,
the image will be written out as MetaIO (.mha) and Fiji launched.  If not,
the function will search for ImageJ (v1).  For ImageJ 1, the image is
written out as Nifti (.nii).

Note that Fiji is ImageJ plus a collection of plugins.  ImageJ with
no I/O plugins does not support MetaIO or Nifti.  ImageJ v1 did come
with a Nifti plugin.  ImageJ v2 has no plugins, while Fiji comes with
MetaIO support.

Change-Id: I53022464d5a1bb994af8de6ad88a4c862fe6651a