COMP: Macros to suppress clang warnings
Clang tends to produce a plethora of warnings.  This patch introduces
3 macros in sitkMacro.h to help the user suppress particular warnings.
The macros sitkClangDiagnosticPush and sitkClangDiagnosticPop save and
restore the state of clang's diagnostic warnings.  sitkClangWarningIgnore
allows the user to suppress any particular warning.

To implement sitkClangWarningIgnore, we use the macro sitkMacroJoin,
which is a pre-processor trick to join two macros, similar boost and
ITK.

The template files in this patch use the sitkClangWarningIgnore macro
to suppress the "-Wunused-local-typedef" warning.

Change-Id: Id724b9c2cdb6b22326e483059f8bccf647ed05f2