ENH: Added debugging message option to Show
Added a boolean parameter, debugOn, to the Show function that will
print out debugging messages.  When enabled, show will print to
stdout 3 things: the search path used to find ImageJ, the full
path to ImageJ, and the full command used to invoke ImageJ.

Also the sitkShowTest enables these messages for the third call
to Show.

Change-Id: I8128196713f826edc1f93532e506f356723abf38