A nicer way of dealing with the PixelIDValue enumeration.
This patch deals with the typelist derived enumeration using
a single function with a large if statement, rather than
a series of functions wrapping individual enum values. The
function is a variant of GetPixelIDValueAsString, but takes
the C++ enumeration names as input. The advantage of this
approach is that it provides a check that all the possible
enumerations are dealt with. This happens because swig generates
a R enumeration containing all names, which are then used
as arguments to the new function. A marker value is returned
when an unrecognised name is passed and a warning is flagged
when the R package is loaded.

Note that there may problems with the various Label options in
GetPixelIDValueAsString, as there is no equality check.

Change-Id: I2699b1af0aa98eb4bf7896aa296a778d6f7205c1