try_compile test for allowing undefined symbols in shared libraries
On Linux (ELF) libraries dynamically loaded by an executable can use
the run-time environment of the executable. This is the recommended
way for Python modules[1] along with other languages loadable
modules.

This patch adds a test to determine if the linker allows unresolved
symbols in shared library, this presumably indicates that the
operating system support loading symbols from the executable into a
loaded module to resolve the module's symbols at run-time.

[1] https://mail.python.org/pipermail/distutils-sig/2016-February/028275.html

Change-Id: I6188d8f74ac1b480906c7fdb1fc715605e2a5af2