Support for building swig from a git repository.
The swig github site is the default repository, and the lastest
release hash is the default tag.

This option requires that automake tools are installed, but there
is currently no check for them in the cmake files.

The option is disabled for MSVC, but this is an assumption on my
part.

Change-Id: Idbb37ae1d3c44d74ca6ce7b289eb937a28ff6a0c