Single functions for pixel access
This patch creates a single SetPixel function and the
corresponding GetPixel function. It isn't quite as nice
as the python equivalent as the swig extend mechanism
doesn't have an obvious way to directly add R code. Instead
we create a dummy C function to force the appropriate
entry in the method list and overwrite it with a hand
generated R function in zA.R

Test now covers both scalars and vectors for setting
and getting. Label image types are also tested.
The Get/Set methods for sitkLabelUInt8 are the
same as for sitkUInt8.

itk::simple::Unused included in dummy Set/Get methods.

This patch is replacing the original, now based on
upstream/master

Change-Id: If3e613c67ee31b3b5b7af566739643103cfe0bfb