BUG: Remove implicit deep copy in ImportImageFilter
The purpose of the ImportImageFilter is to enable SimpleITK to share
the buffer with an external application. The shallow copy or sharing
of the buffer with SimpleITK is now restored.

The implicit copy was occurring because the constructed ITK image was
used by the sitk::Image. Then the sitk::Image was modified, causing
the copy. This has been corrected by setting all attributes in the ITK
image, and constructing the sitk::Image on the return.

Change-Id: Iebf0df2c4a81fb6db6f09049e36fd58759f451d2