New image tests for Java
Initial checkin of a new test for Java.  The Java.ImageTests has two
simple tests.  The first one, BasicImageTest, creates a small image,
sets a pixel and gets it back.

The second test, LabelShapeStatisticsTest, creates a small image with
a smaller square label in the middle.  Then the
LabelShapeStatisticsImageFilter is applied to it and the results
are checked.  It tests for the number of labels (1), the number of
pixels for each label (16 for label 1) and gets the perimeter
of that label.

Change-Id: I583a6285319b8b4ca8ba822f8d4484bef2775b97