Fix for CMake's semicolon issue on Windows
The sitk_add_java_test function needs to pass a CLASSPATH for Java
down to sitk_add_tests.  CLASSPATH contains two directories that are
separated by a semicolon on Windows.  However CMake uses semicolons
to separated elements in a string into a list.  And in CMake function
arguements are passed as a single string, separated by semicolons.

So to get a CLASSPATH string such as "path1;path2" to not get separated
into list elements by CMake, we need to use backslashes to escape
the semicolon.  And in this case, we need our CLASSPATH to make it
down through 4 levels of function calls.  So we have to use SEVEN
backslashes: "path1\\\\\\\;path2".

Change-Id: Idb1c8f94db04791986eba1eea94ecfc924edaa37