Advances in bracket operator/slicing
The R bracket operator now uses the Slice and Extract filters,
rather than custom C code. The major change is that arbitary
slicing is no longer supported - there must be consistent
spacing along each extracted dimension. The R code
checks this kind of issue and attempts to produce
sensible warnings/errors.

A notable difference to array indexing is:

a[c(),1:2] or a[0,1:2]

These will produce an empty array with the first dimension size of 0.
This doesn't have any particular use, and doesn't make sense for images,
so this operation will raise an error for images.

The Method code is quite complicated in terms due to the potentially
arbitary number of dimensions that may be used. I have what I think is
a solution, but will continue to search for improvments in dealing with
the scenario.

There is a test that covers origins, spacings and a few warning/error
scenarios.

A manual doc file has been created to cover some of the basics of the
image class:

library(SimpleITK)
help("SimpleITK_Image-class")

Change-Id: I2dcfcc042154f910360b1c1fac2ae037a10736d6