Build a loadable Lua module
Previously we only built a standalone Lua executable.  Now we also
build a SimpleITK module that can be loaded into any Lua interpreter.

This set of patches also changes the Lua tests to use the system
Lua executable and load SimpleITK as a module in that executable.

Also building the standalone Lua executable is now an option that
defaults to OFF.