Adding circle.yml for continuous integration with CircleCi
This test configuration used distcc to take advantages of multiple
containers on CircleCi. The .distcc/hosts file has is automatically
generated for the configured number of containers. Tweaks have been with
"--randomize" to use all containers evenly ( this seems to help keep them
alive, and not terminate the build ), and to slightly over subscribe
them with 3.

The ITK git download take a couple minutes, so it is pre-downloaded
and saved in the cache.

The ctest tests are split among the containers.

The ITK SuperBuild is done in one setup, then the full build this is
split due to the amount of text output the ITK build and install fill
the web display buffer along. By splitting into two commands, the text
output is better displayed.

Change-Id: Id864cbd8c2598412300d9c5e0a648259f0f5b392