Improvements to Rd file generation
The usages for the procedural interface have been combined into
a single usage section - the previous strategy of multiple
usage sections is not legal in R documentation.

The class documentation is now based around the recommended style
for S4 classes. This is not a perfect match for the wrapping style,
but better than previously. It now has a methods section which
is used to document c++ methods.

There are very few errors/warnings when creating the package. Those
remaining - AntiAliasBinaryImageFilter - come from the latex maths
messing up the markup.

Change-Id: Iabd221e2c3a55dae4454113e7089a2a0a4bcf9b3