Support sitkCompareDriver for MD5 and RMS image compare
Adding to the sitkCompareDriver the "--compare" and the
"--compare-MD5" options to compare images. This is done in connection
with adding the same arguments to the sitk_add_test CMake
function. This enable the language test to produce output and have the
comparison of the results done with sitkCompareDriver, and have the
options specified in CMake ( possibly generated ).