Move Rdefines and Rversions outside the extern C.
Writing R extensions says:

"Most R header files can be included within C++ programs but they
should not be included within an extern "C" block
(as they include system headers"

This patch includes the modified rrun.swg file that I've submitted
to the swig project. It will be used instead of the installed version.
The version shipped with swig generates the runtime C/C++ code with
an extern C block around Rdefines.h. On most platforms this hasn't
been an issue, but the swig generated wrapper code fails to compile
with recent combinations of R and Xcode on OSX.

Change-Id: I9c00d8af085b4ed258d958cb17bbc2c4fc031031