Use CMake ExternalData module to download ExternalProject source
Using the ExternalData module to download the source code in the
SuperBuild project enables local caching of the download in the build
tree, and on the local system with the "ExternalData_OBJECT_STORES"
environment variable. This may help improve the performance of
continuous integration services.

Additionally, the ExternalData module enables multiple download
sources for redundancy and download reliability.

Change-Id: Idb4f99d3a696ad12bd70a415966850f0c9381a43