Add support for scalar operations on vectors
Add support for multiplying a vector image by a constant scalar. This
patch changes the itk filter instantiated for all filters which use
the BinaryFunctorFilter template, from just the same binary type
for both images and scalars, to using a "ValueType" scalar image and
image type. This is the same as the pixel type for scalar images. For
VectorImages it is the same as the component type and for complex
images it is the same as the value_type.

With the new BinaryGeneratorImageFilter, the instantiate of additional
permutations of vector and scalars is bigger but less significant the
the BinaryImageFilter approach.

WIP: The combination if Vector Images and scalar image is being
 instantiated, for many filters but it is not being exposed.

Restores features that were provided by the MultipleByConstanst and
similarly named filters, before they were deprecated and removed.

Change-Id: I68143cf2a03d160136c030d5c7781d1f4db47308