Propogate C++ exceptions into R.
The swig bindings deal with exceptions in a range of ways.
Exceptions generated in the binding code become R exceptions,
but most exceptions generated by the wrapped code become warnings.

These patches check the type of objects from a few types of call
and raise an R exception if the return type is not what is expected.

Change-Id: I31101bd520d4d6b00a196885b585c1d0c3f9cf50