Initial checkin of GenerateDoc.py
GenerateDoc.py is a re-write, in Python, of the
GenerateDocumentation.groovy script.  I wrote this so that
we would no longer need groovy to update SimpleITK's documentation.

The script works the same as the groovy version:
  GenerateDoc.py SimpleITKClass.json ITK_XML_PATH

The script will update, in place, a SimpleITK class's JSON file.
It finds the corresponding ITK class's XML file, traverse the XML
structure and pulls any method descriptions that it finds from the XML
to the JSON.

Change-Id: I4129d7e6c59302f501bde2c329089238645bf794