Add advanced CMake variable to determine default value of WRAP_*
This new variable enables used to turn off all wrapping by
default. When WRAP_DEFAULT is on, then if a language is detected on
the system it will be enabled for the initial configuration of
SimpleITK. If this option is disabled it will all wrapped languages
will be disabled by default.

Change-Id: I8d0b75533845f65d1c1441d07252b7e823c1971c