Add support for GTEST_ROOT as source directory.
The current recommendation for the Google Test project is to build the
library as part of the project with the add_subdirectory
command. While Redhat based Linux distribution contain binary packages
with the library Debian based systems only include the source code.

This patch adds the feature to FindGTest to enable usage of  GTest
source code by adding the source directory to a project.

Change-Id: Ie0cfb1fcbfbf6c7854a793c0e094c01fbb3768cd