Create BasicFilter libraries corresponding to ITK modules
Previously, the list of BasicFilters was simply split based on a fixed
stride. Now, with the added itk_module field in the json files, each
ITK module has a corresponding SimpleITK library. This reduces the
size of the libraries since ITK module tend to reuse the same base
classes, such and the morphology directories.