Set target properties for using ITK modules instead of including use file
Created a new CMake function "sitk_target_use_itk", which configures a
target's properties to use a list of ITK modules. This can be uses
instead of including the ITK_USE_FILE. The target specific properties
can be more specific to the ITK modules needed, reducing the number of
included directories and linking libraries thereby improving
compilation.

As the ITK_USE_FILE is not included, it is not longer needed to
explicitly disable the factory registration in each module. Only the
ITK IO modules does the factory registration for ImageIO, and the
TransformIOs.