BUG: removed the loadPrivateTags parameter from GetGDCMSeriesFileNames.
This parameter was not relevant for users, we always want to skip
reading the private tags when getting the series file names. Also got
rid of the separate functions in which one received a series and the
other did not (added a default value of empty string as the series).

Change-Id: Ie9f7d6119b26b8f1a1681199c9c035306eb48b57