Improvements to garbage collection in R package
itk::simple::Images lead to leaks in the R package.
Swig attaches finalizers to the results
of constructor calls, but not to objects returned by method calls.

The change here isn't as general as I'd like, but cures the most
dramatic problems. These are illustrated by the code below. I'm
not sure that the code can be included as a useful test, as
it requires something external to check process RAM usage.

library(SimpleITK)

InputIm <- Image(512, 512, 512, 'sitkFloat64')
ProcessedIms <- replicate(5, ShiftScale(InputIm, 1, 2))
ProcessedIms <- 0
gc()

Change-Id: Iacd3e9d40048e3ed5f8ea57cfd967e56e89205e4