Use find_package for install and build path for build for ITK
The ITK targets file was previously hard coded to the path during
SimpleITK build. This does not allow a relocated ITK install, and is
problematic when ITK is build in the SimpleITK superbuild.

This is now corrected to only use the build path when SimpleITK is in
the build tree. When SimpleITK is installed it now uses find_package
to locate the proper version of ITK.

ITK is required for its library targets.