Set default ITK modules configuration as override-able variables
Use the mechanism which passes ITK/Module_ITK prefixed CMake variables
to the external ITK project. Implemented for the Review module and the
build default modules options.  This allows the user to disable
arbitrary ITK modules in the SimpleITK Superbuild, and have SimpleITK
not wrap them to create a smaller ( faster build ) of SimpleITK.