Add correct alternative baseline image
ITK 4.12 contains updates a fix in the IsoContourFilter which correct
the results of the ApproximateSignedDistanceMapImageFilter with values
of 1, 0.

This is from the ITK commit: 0ad648d9fd15293236391f2206c45e04af2d394b

The old baseline should be removed in the future, but we will keep old
incorrect one to enables test with old version of ITK to pass.