Add MetaDataDictionaryArrayUpdate parameter to ImageSeriesReader
This allows for the reading and creation of a MetaDataDictionary for
each slice in the series reader. It defaults to false to minimize
resources used. It does not need to be added to the procedural
interface since it will only be useful with active measurements
available in the object's interface.