BUG: remove package_data hack for overloading build_ext
The prior hack to include the _SimpleITK.so python library,
incorrectly marked the package as a purelib. This is now causing
problems with audit wheel preventing generation of many_linux
binaries.

We now overload the "build_ext" command to provide a different build
step which simply copies the specified python library file to the
expected output directory for setuputils or disutils to manage.

Use full path to Python extension library.