Do not create /.git folder when the project is configured
A .git folder kept popping up on in the root folder of my and many of my coworkers' computers.
It was really annoying, because it confused git clients to think that all folder is under revision control
and we had to delete the folder manually. After that the .git folder appeared again, and again...
It remained a mystery for several years ... until now.

We've finally found the culprit: it was a copy-paste error in sitkCheckSourceTree.cmake.
This commit fixes the issue.