Update README
Include information about how new users can contribute to SimpleITK. Two references are given that explain how the Git workflow process works and how to setup your computer to start contributing to a Github project.

Provide a description of the three primary branches in SimpleITK in a clean, organized way.

Include a documentation section that explains the various levels of documentation: Doxygen, ReadTheDocs, Wiki, and Notebooks. Made it clear that the Wiki is obsolete and is being replaced by ReadTheDocs.