Use sitkDebugMacro over std::cout for debugging information
Introducing the sitkDebugMacro which prints out detailed information
when enabled globally or at the object level. Unlike the ITK version,
even in release mode the information will be printed.

Remove the usage of directly printing to std::cout. Instead use the
ITK debug formatting, and print to itk OutputWindow interface.