Add meta-data information interface to the ImageFileReader
This new interface available only to the ImageFileReader's object
oriented interface provides access to the image information and the
meta-data dictionary provided by the itk ImageIO after
ImageIOBase::ReadImageInformation is called. This provides and
interface to obtain the size of a file and it's meta-data before the
bulk data is loaded.