Large refactor of GenerateDoc.py
The script now deletes any simplesect node that has a title
node with 'Wiki Examples' text.

This search for Wiki Examples and removal is performed in the
process_xml function.  The XML package used was changed from
ElementTree to lxml so process_xml could use lxml's more powerful
xpath search.

Also the code has been reorganized to move the main execution code
down to the bottom main section and the code to find the XML file
was pulled out to a function.

And finally the formatting of the code was changed to be more compliant
with the PEP 8 style for Python.  'autopep8' was run on the code,
and the case style of variables and functions was changed.