Adding request of IO region to ImageFileReader
Adding to the ImageFileReader, set methods to request a region to be
read. The behavior corresponds to the ExtractImageFilter, where any
dimension with a size of "0" is considered a dropped dimension, and
the output dimension will be reduce. Additionally this will enable
"streamed" reading of large image files if the ImageIO supports the
feature.

Internally this is implemented by always reading into a 5D image, and
then using the Extract filter to reduce.