Remove local CMake GTest module, use gtest_discover_tests
The CMake versions of FindGTest, and GoogleTest are sufficiently up to
date for our requirements so we don't need a local version. We have
replaced the slow configuration time introspection with build time
introspection to determine which Google tests are available. This
eliminate the time intensive CMake configuration when a Google test
files is updated.