Changed the HTML template of the documentation
Changed the HTML template to sphinx_rtd_theme and added style.css file
and breadcrumbs.html to modify the existing layout, features, and
readability. Also modified the appropriate files to remove on-page toc's
and only navigate using the toc on the left sidebar. Also updated
style.css to allow text wrapping in tables