Modified generation of namespace file
The previous namespace file exports all functions in the package,
making them available to the user. This includes the many overloaded
methods. In reality, none of these methods are called directly
by the user, as all access is via the $ method, for example
filterobj$filtermethod(). Thus there is no need to export the filtermethod
function, only the $ method. This patch uses the NAMESPACE file generated
by swig as the basis for automatically creating a more conservative
set of exported names. The swig namespace isn't appropriate for using
directly, but is a convenient list of names.

There are limitations associated with static methods. There aren't
many of these and they are currently listed manually in the mkNamespace.R
file.