Initial checkin of the sitkImageViewer class
This is a new class for displaying images in SimpleITK.  It
takes the functionality of the Show function and encapsulates it
in a C++ class.  It exposes various parameters that Show was using
to find the viewing application, via methods of the ImageViewer
class.  This gives the user more freedom and flexibility in
specifying a viewing application.

The one change in functionality from the previous Show function,
is that the environment variables (SITK_SHOW_COMMAND,
  SITK_SHOW_3D_COMMAND, SITK_SHOW_COLOR_COMMAND, and
  SITK_SHOW_EXTENSION) are only checked at the launch of the SimpleITK
process.  Previously Show would check these variables every time the
function was called.